Vendor offer

+++ back-end/fitnessfinder/src/main/java/com/project/fitnessfinder/domain/entity/database/Address.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class Address {
 
     private String fullAddress;
-    private Long latitude;
-    private Long longitude;
+    private double latitude;
+    private double longitude;
 }
